requirements.txt
fastmcp>=2.10.0
aiohttp>=3.12.0
pydantic>=2.11.0
python-dotenv>=1.1.0
